Correct Dockerfile to take account of build context
The previous approach of copying python code into docker container cannot work, since that is now
outside of the build context. Instead, the Pyabolism module code is cloned from GitHub. Module is now
also installed in 'develop' mode, allowing the directory to be remapped on creation of the container, and
updates to the code being immediately reflected in the Python run time.