Add prefix to variable and constraint names in Gurobi LP
Previously variables and constraints were named using just the reaction and metabolite IDs,
but for more sophisticated algorithms it is useful to have unique naming. For example, in
situations where multiple models might be combined into a single LP. By default a random
string is assigned to the ModelName attribute of the Gurobi model, and this is used as a prefix
for all variable and constraint names.